Centralize materialized view generation
This commit splits materialized view generation methods into individual
modules that are flattened in "dal-materialized-views". This change
helps authors not have to keep track of the crate structure, and keeps
it flattened to mimic edda's flattened usage of materialized views.

Another benefit of this change is that all MVs assembled are now in
"dal-materialized-views" and NOT in "dal" or "sdf-server". This should
help build times for the latter two crates and is now solely-scoped to
the "edda" build chain.

Domain nesting (e.g. "component" contains all component methods and
"aciton" contains its methods and nests the "prototype" methods in a
sub-module) causes needless segmentation of the crate. In addition to
that change, every MV method is called "assemble" and is instrumented
at the debug level. This standardization should help with local and
spot-debugging in limited production scenarios.

Bonus change: getting the reference ID for a geometry now returns an
option rather than an unused error.
