+++ lib/dal/tests/integration_test/mod.rs
@@ -14,10 +14,9 @@ mod frame;
 mod func;
 mod input_sources;
 mod management;
-mod materialized_view;
+mod materialized_views;
 mod migrate;
 mod module;
-mod newhotness;
 mod node_weight;
 mod pkg;
 mod prompt_overrides;
